Level._content, tooltip, in_game
- change structure of _content. Now it's 3D array, where 3'rd d. is an array
with dictionaries {object:reference}.
- add content_has(x,y,?) function, which replaces _content[x][y].has(?) call
(it doesn't work with new structure of _content)
- replace old content_has calls.
- move responsibility of adding to _content to entities instead of MapGen's.

- add "in game" state which actually shows tooltip with _content's tile's
content.
- tooltip fixes

A change of _content and inplement content_has was necessary i we want have
easy access to both names of entities and they propertys/methods. The downside
of content_has() is that it needs (less than) one iteration in tile, and it's
propably slower than build-in Array.has() function.